requirements.txt
dotenv=0.9.9
crewai==0.98.0
crewai-tools==0.32.1
langgraph==0.2.66
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-core==2.24.0
google-api-python-client==2.159.0
grandalf
pyppeteer
mermaid
pygraphviz
llama-index-core==0.12.23.post2
llama-index-readers-file==0.4.6
lama-index-llms-openai==0.3.25
llama-index-embeddings-openai==0.3.1
langchain_milvus--0.1.8
pymilvus==2.5.5
pymupdf
lxml==5.3.1
llama-index-vector-stores-milvus==0.5.0
#pip install llama-index[all]

langchain-ollama==0.3.0

####
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.67.1
grpcio-health-checking==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1734788189411/work
ipywidgets==8.1.5
isoduration==20.11.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
json5==0.10.0
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257272359/work
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kubernetes==31.0.0
lancedb==0.18.0
langchain==0.3.15
langchain-cohere==0.3.5
langchain-community==0.3.15
langchain-core==0.3.31
langchain-experimental==0.3.4
langchain-milvus==0.1.8
langchain-openai==0.2.14
langchain-text-splitters==0.3.5

langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.147
langtrace-python-sdk==3.6.1
linkup-sdk==0.2.2
litellm==1.57.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mem0ai==0.1.47
mistune==3.1.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
patronus==0.0.17
pdfminer.six==20231228
pdfplumber==0.11.5
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
portalocker==2.10.1
posthog==3.9.2
prometheus_client==0.21.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.0.0
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyJWT==2.10.1
pylance==0.22.0
pymilvus==2.5.4
pyOpenSSL==24.3.0
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.392.post0
pysbd==0.3.4
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
python-json-logger==3.2.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1666828590571/work
qdrant-client==1.13.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
schema==0.7.7
scrapegraph_py==1.10.2
selenium==4.28.0
Send2Trash==1.8.3
sentry-sdk==2.20.0
serpapi==0.1.5
shapely==2.0.6
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
snowflake==1.0.2
snowflake-connector-python==3.12.4
snowflake._legacy==1.0.0
snowflake.core==1.0.2
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.1.26
SQLAlchemy==2.0.37
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.45.2
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
tornado @ file:///D:/bld/tornado_1666788744359/work
tqdm==4.67.1
trace-attributes==7.2.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.48.2
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.23
uvicorn==0.34.0
validators==0.34.0
watchfiles==1.0.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
weaviate-client==4.10.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
